requirements.txt
faiss-cpu
sentence-transformers
pandas
numpy
openai-whisper
pyttsx3
google-generativeai
python-dotenv
torch
fastapi
uvicorn
soundfile

